+++ atproto/repo/mst/mst.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
-	"slices"
 
 	"github.com/ipfs/go-cid"
 )
@@ -52,6 +51,19 @@ type NodeEntry struct {
 
 func (n *Node) IsEmpty() bool {
 	return len(n.Entries) == 0
+}
+
+// Checks if the sub-tree (this node, or any children, recursively) contains any CID references to nodes which are not present.
+func (n *Node) IsPartial() bool {
+	for _, e := range n.Entries {
+		if e.ChildCID != nil && e.Child == nil {
+			return true
+		}
+		if e.Child != nil && e.Child.IsPartial() {
+			return true
+		}
+	}
+	return false
 }
 
 // Returns true if this entry is a key/value at the current node
@@ -157,87 +169,6 @@ func findInsertionIndex(n *Node, key []byte) (idx int, split bool, retErr error)
 	return len(n.Entries), false, nil
 }
 
-// Adds a key/CID entry to a sub-tree defined by a Node. If a previous value existed, returns it.
-//
-// If the insert is a no-op (the key already existed with exact value), then the operation is a no-op, the tree is not marked dirty, and the val is returned as the 'prev' value.
-//
-// n: Node at top of sub-tree to operate on
-// key: key or path being inserted. must not be empty/nil
-// val: CID value being inserted
-// height: tree height to insert at, derived from key. if a negative value is provided, will be computed; use -1 instead of 0 if height is not known
-func Insert(n *Node, key []byte, val cid.Cid, height int) (*Node, *cid.Cid, error) {
-	if height < 0 {
-		height = HeightForKey(key)
-	}
-
-	if n == nil {
-		return nil, nil, fmt.Errorf("operating on nil tree/node")
-	}
-
-	for height > n.Height {
-		// if the new key is higher in the tree; will need to add a parent node, which may involve splitting this current node
-		return insertParent(n, key, val, height)
-	}
-
-	// if key is lower on the tree, we need to descend first
-	if height < n.Height {
-		return insertChild(n, key, val, height)
-	}
-
-	// look for existing key
-	idx := findExistingEntry(n, key)
-	if idx >= 0 {
-		e := n.Entries[idx]
-		if *e.Value == val {
-			// same value already exists; no-op
-			return n, &val, nil
-		}
-		// update operation
-		prev := e.Value
-		n.Entries[idx].Value = &val
-		n.Dirty = true
-		return n, prev, nil
-	}
-
-	// insert new entry to this node
-	idx, split, err := findInsertionIndex(n, key)
-	if err != nil {
-		return nil, nil, err
-	}
-	n.Dirty = true
-	newEntry := NodeEntry{
-		Key:   key,
-		Value: &val,
-	}
-
-	if !split {
-		// TODO: is this really necessary? or can we just slices.Insert beyond the end of a slice?
-		if idx >= len(n.Entries) {
-			n.Entries = append(n.Entries, newEntry)
-		} else {
-			n.Entries = slices.Insert(n.Entries, idx, newEntry)
-		}
-		return n, nil, nil
-	}
-
-	// we need to split
-	e := n.Entries[idx]
-	left, right, err := splitNode(e.Child, key)
-	if err != nil {
-		return nil, nil, err
-	}
-	// remove the existing entry, and replace with three new entries
-	n.Entries = slices.Delete(n.Entries, idx, idx+1)
-	n.Entries = slices.Insert(
-		n.Entries,
-		idx,
-		NodeEntry{Child: left},
-		newEntry,
-		NodeEntry{Child: right},
-	)
-	return n, nil, nil
-}
-
 // returns the lowest key along "left edge" of sub-tree
 func getLowestKey(n *Node) []byte {
 	// TODO: make this a method on Node, and return an error (?)
@@ -277,322 +208,6 @@ func getHighestKey(n *Node) []byte {
 	}
 	// TODO: throw error on this case?
 	return nil
-}
-
-func splitNodeEntries(n *Node, idx int) (*Node, *Node, error) {
-	if idx == 0 || idx >= len(n.Entries) {
-		return nil, nil, fmt.Errorf("splitting at one end or the other of entries")
-	}
-	left := Node{
-		Height:  n.Height,
-		Dirty:   true,
-		Entries: n.Entries[:idx],
-	}
-	right := Node{
-		Height: n.Height,
-		Dirty:  true,
-		// don't use the same slice here
-		Entries: append([]NodeEntry{}, n.Entries[idx:]...),
-	}
-	if left.IsEmpty() || right.IsEmpty() {
-		return nil, nil, fmt.Errorf("one of the legs is empty (idx=%d, len=%d)", idx, len(n.Entries))
-	}
-	return &left, &right, nil
-}
-
-func splitNode(n *Node, key []byte) (*Node, *Node, error) {
-	if n.IsEmpty() {
-		// TODO: this feels defensive and could be removed
-		return nil, nil, fmt.Errorf("tried to split an empty node")
-	}
-
-	idx, split, err := findInsertionIndex(n, key)
-	if err != nil {
-		return nil, nil, err
-	}
-	if !split {
-		// simple split based on values
-		return splitNodeEntries(n, idx)
-	}
-
-	// need to split recursively
-	e := n.Entries[idx]
-	lowerLeft, lowerRight, err := splitNode(e.Child, key)
-	if err != nil {
-		return nil, nil, err
-	}
-	left := &Node{
-		Height:  n.Height,
-		Dirty:   true,
-		Entries: []NodeEntry{},
-	}
-	left.Entries = append(left.Entries, n.Entries[:idx]...)
-	left.Entries = append(left.Entries, NodeEntry{Child: lowerLeft})
-	right := &Node{
-		Height:  n.Height,
-		Dirty:   true,
-		Entries: []NodeEntry{NodeEntry{Child: lowerRight}},
-	}
-	if idx+1 < len(n.Entries) {
-		right.Entries = append(right.Entries, n.Entries[idx+1:]...)
-	}
-	return left, right, nil
-}
-
-// inserts a node "above" this node in tree, possibly splitting the current node
-func insertParent(n *Node, key []byte, val cid.Cid, height int) (*Node, *cid.Cid, error) {
-	var parent *Node
-	if n.IsEmpty() {
-		// if current node is empty, just replace directly with current height
-		parent = &Node{
-			Height: height,
-			Dirty:  true,
-		}
-	} else {
-		// otherwise push a layer and recurse
-		parent = &Node{
-			Height: n.Height + 1,
-			Dirty:  true,
-			Entries: []NodeEntry{NodeEntry{
-				Child: n,
-			}},
-		}
-	}
-	// regular insertion will handle any necessary "split"
-	return Insert(parent, key, val, height)
-}
-
-// inserts a node "below" this node in tree; either creating a new child entry or re-using an existing one
-func insertChild(n *Node, key []byte, val cid.Cid, height int) (*Node, *cid.Cid, error) {
-	// look for an existing child node which encompases the key, and use that
-	idx := findExistingChild(n, key)
-	if idx >= 0 {
-		e := n.Entries[idx]
-		if e.Child == nil {
-			return nil, nil, fmt.Errorf("could not insert key: %w", ErrPartialTree)
-		}
-		newChild, prev, err := Insert(e.Child, key, val, height)
-		if err != nil {
-			return nil, nil, err
-		}
-		if prev != nil && *prev == val {
-			// no-op
-			return n, &val, nil
-		}
-		n.Dirty = true
-		n.Entries[idx].Child = newChild
-		return n, prev, nil
-	}
-
-	// insert a new child node. this might be recursive if the child is not a *direct* child
-	idx, split, err := findInsertionIndex(n, key)
-	if err != nil {
-		return nil, nil, err
-	}
-	if split {
-		return nil, nil, fmt.Errorf("unexpected split when inserting child")
-	}
-	n.Dirty = true
-	newChild := &Node{
-		Height: n.Height - 1,
-		Dirty:  true,
-	}
-	newChild, _, err = Insert(newChild, key, val, height)
-	if err != nil {
-		return nil, nil, err
-	}
-	newEntry := NodeEntry{
-		Child: newChild,
-	}
-	if idx == len(n.Entries) {
-		n.Entries = append(n.Entries, newEntry)
-	} else {
-		n.Entries = slices.Insert(n.Entries, idx, newEntry)
-	}
-	return n, nil, nil
-}
-
-// Removes key/value from the sub-tree provided, returning a new tree, and the previous CID value. If key is not found, returns unmodified subtree, and nil for the returned CID.
-//
-// n: Node at top of sub-tree to operate on. Must not be nil.
-// key: key or path being inserted. must not be empty/nil
-// height: tree height corresponding to key. if a negative value is provided, will be computed; use -1 instead of 0 if height is not known
-func Remove(n *Node, key []byte, height int) (*Node, *cid.Cid, error) {
-	// TODO: better handling of "is this the top"
-	top := false
-	if height < 0 {
-		top = true
-		height = HeightForKey(key)
-	}
-
-	if height > n.Height {
-		// removing a key from a higher layer; key was not in tree
-		return n, nil, nil
-	}
-
-	if height < n.Height {
-		// XXX: handle case of this returning an empty node; in which case reset height to empty tree?
-		return removeChild(n, key, height)
-	}
-
-	// look at this level
-	idx := findExistingEntry(n, key)
-	if idx < 0 {
-		// key not found
-		return n, nil, nil
-	}
-
-	// found it! will remove from list
-	n.Dirty = true
-	prev := n.Entries[idx].Value
-
-	// check if we need to "merge" adjacent nodes
-	if idx > 0 && idx+1 < len(n.Entries) && n.Entries[idx-1].IsChild() && n.Entries[idx+1].IsChild() {
-		fmt.Println("MERGE") // XXX
-		if n.Entries[idx-1].Child == nil || n.Entries[idx+1].Child == nil {
-			return nil, nil, fmt.Errorf("can't merge partial nodes")
-		}
-		newChild, err := mergeNodes(n.Entries[idx-1].Child, n.Entries[idx+1].Child)
-		if err != nil {
-			return nil, nil, err
-		}
-		// XXX:
-		if err := DebugSiblingChild(newChild); err != nil {
-			fmt.Println("BAD CHILD MERGE")
-		}
-		n.Entries = slices.Delete(n.Entries, idx, idx+2)
-		n.Entries[idx-1] = NodeEntry{Child: newChild}
-	} else {
-		// simple removal
-		n.Entries = slices.Delete(n.Entries, idx, idx+1)
-	}
-	// XXX
-	if err := DebugSiblingChild(n); err != nil {
-		fmt.Println("BAD MERGE")
-	}
-
-	// check if top of node is now just a pointer
-	if top {
-		for {
-			if len(n.Entries) != 1 || !n.Entries[0].IsChild() {
-				break
-			}
-			if n.Entries[0].Child == nil {
-				return nil, nil, fmt.Errorf("can't prune partial tree") // TODO: wrap partial
-			}
-			n = n.Entries[0].Child
-		}
-	}
-	return n, prev, nil
-}
-
-func mergeNodes(left *Node, right *Node) (*Node, error) {
-	idx := len(left.Entries)
-	n := &Node{
-		Height:  left.Height,
-		Dirty:   true,
-		Entries: append(left.Entries, right.Entries...),
-	}
-	if n.Entries[idx-1].IsChild() && n.Entries[idx].IsChild() {
-		// need to merge recusively
-		lowerLeft := n.Entries[idx-1]
-		lowerRight := n.Entries[idx]
-		if lowerLeft.Child == nil || lowerRight.Child == nil {
-			return nil, fmt.Errorf("can't merge partial child nodes") // TODO: wrap partial error
-		}
-		lowerMerged, err := mergeNodes(lowerLeft.Child, lowerRight.Child)
-		if err != nil {
-			return nil, err
-		}
-		n.Entries[idx-1] = NodeEntry{Child: lowerMerged}
-		n.Entries = slices.Delete(n.Entries, idx, idx+1)
-	}
-	return n, nil
-}
-
-// internal helper
-func removeChild(n *Node, key []byte, height int) (*Node, *cid.Cid, error) {
-	// look for a child
-	idx := findExistingChild(n, key)
-	if idx < 0 {
-		// no child pointer; key not in tree
-		return n, nil, nil
-	}
-
-	e := n.Entries[idx]
-	if e.Child == nil {
-		// partial node, can't recurse
-		return nil, nil, fmt.Errorf("could not remove key: %w", ErrPartialTree)
-	}
-	newChild, prev, err := Remove(e.Child, key, height)
-	if err != nil {
-		return nil, nil, err
-	}
-	if prev == nil {
-		// no-op
-		return n, nil, nil
-	}
-
-	// if the child node was updated, but still exists, just update pointer
-	if !newChild.IsEmpty() {
-		n.Dirty = true
-		n.Entries[idx].Child = newChild
-		return n, prev, nil
-	}
-
-	// if new child was empty, remove it from entry list; note that *this* entry might now be empty
-	n.Dirty = true
-	n.Entries = slices.Delete(n.Entries, idx, idx+1)
-	return n, prev, nil
-}
-
-// Reads the value (CID) corresponding to the key. If key is not in the tree, returns (nil, nil).
-//
-// n: Node at top of sub-tree to operate on. Must not be nil.
-// key: key or path being inserted. must not be empty/nil
-// height: tree height corresponding to key. if a negative value is provided, will be computed; use -1 instead of 0 if height is not known
-func Get(n *Node, key []byte, height int) (*cid.Cid, error) {
-	if height < 0 {
-		height = HeightForKey(key)
-	}
-
-	if height > n.Height {
-		// key from a higher layer; key was not in tree
-		return nil, nil
-	}
-
-	if height < n.Height {
-		// look for a child node
-		idx := findExistingChild(n, key)
-		if idx >= 0 {
-			if n.Entries[idx].Child == nil {
-				return nil, fmt.Errorf("could not search for key: %w", ErrPartialTree)
-			}
-			return Get(n.Entries[idx].Child, key, height)
-		}
-	}
-
-	// search at this height
-	idx := findExistingEntry(n, key)
-	if idx >= 0 {
-		return n.Entries[idx].Value, nil
-	}
-
-	// not found
-	return nil, nil
-}
-
-// Checks if the sub-tree (this node, or any children, recursively) contains any CID references to nodes which are not present.
-func (n *Node) IsPartial() bool {
-	for _, e := range n.Entries {
-		if e.ChildCID != nil && e.Child == nil {
-			return true
-		}
-		if e.Child != nil && e.Child.IsPartial() {
-			return true
-		}
-	}
-	return false
 }
 
 func NewTreeFromMap(m map[string]cid.Cid) (*Node, error) {
